Adds a reference to the Newtonsoft JSON parser which is used by the Scoreboard.cs class for parsing the JSON scores.
Fetching scores is nearly complete, some testing is required before I finish this. The other score methods are being worked on now.

Fixed a bug in keypair that would include enclosing quotes around the value.

Added some eyecandy to SilverJolt in the form of a number of colored cubes as well as added some new features and fixes.